+++ src/components/patterns/carbon-calculator/components/__tests__/CarbonCalculatorResults.jest.spec.js
@@ -9,7 +9,7 @@ config.global.renderStubDefaultSlot = true;
 const results = 'Results';
 const kgsOf = 'KGs of CO2';
 const comparison = 'That\'s the equivalent of xxx';
-const kgsPerDay = 'KGs of CO2 per day';
+const kgsPerDay = 'xxx KGs of CO2 per day';
 const perDaySuccess = 'Congratulations!';
 const perDayTarget = 5;
 
@@ -56,7 +56,30 @@ describe('VsCarbonCalculatorResults', () => {
         it('should correctly render the labelsMap `kgsPerDay`', () => {
             const wrapper = factoryShallowMount();
 
-            expect(wrapper.text()).toContain(kgsPerDay);
+            expect(wrapper.text()).toContain('KGs of CO2 per day');
+        });
+
+        it('should render the total kilos value', () => {
+            const wrapper = factoryShallowMount({
+                totalKilos: 10,
+                stayDuration: 4,
+            });
+
+            const perDay = wrapper.find('.vs-carbon-calculator-results__per-day');
+
+            expect(perDay.text()).toContain('2.5');
+        });
+
+        it('should render the total kilos value with a comma delimeter if the locale is set to de-de', () => {
+            const wrapper = factoryShallowMount({
+                totalKilos: 10,
+                stayDuration: 4,
+                language: 'de-de',
+            });
+
+            const perDay = wrapper.find('.vs-carbon-calculator-results__per-day');
+
+            expect(perDay.text()).toContain('2,5');
         });
 
         it('should not render the labelsMap `perDaySuccess` if perDay emissions are above the `perDayTarget` value', () => {